refactor: шаблон бота в отдельный компонент
