+++ src/components/dashbord/bots-templates/bots-templates.tsx
@@ -2,51 +2,12 @@ import { FC, useEffect, useRef, useState } from 'react';
 import { useDraggable } from 'react-use-draggable-scroll';
 import styles from './bots-templates.module.scss';
 
-import ButtonAddSampleBot from '../../../ui/buttons/button-add-sample-bot/button-add-sample-bot';
-import BotTemplatePopup from '../../popups/bot-template-popup/bot-template-popup';
-import ModalPopup from '../../popups/modal-popup/modal-popup';
-import useModal from '../../../services/hooks/use-modal';
 import { useAppDispatch, useAppSelector } from '../../../services/hooks/hooks';
 import { getTemplatesBotsAction } from '../../../services/actions/bots/getTemplatesBots';
 import { getAccessToken } from '../../../auth/authService';
 import { getTemplatesBotsSel } from '../../../utils/selectorData';
 import Typography from '../../../ui/typography/typography';
-import { TTemplateBot } from '../../../services/types/bot';
-
-const Template: FC<{ template: TTemplateBot }> = ({ template }) => {
-  const importImage = async () => {
-    try {
-      const imageModule = await import(
-        `../../../images/icon/template/${template.icon}.svg`
-      );
-      return imageModule.default;
-    } catch (error) {
-      return 'null';
-    }
-  };
-
-  const [image, setImage] = useState<string>('');
-  const { isModalOpen, closeModal, openModal } = useModal();
-
-  useEffect(() => {
-    importImage().then((importedImage) => {
-      setImage(importedImage);
-    });
-  }, [template.icon]);
-
-  return (
-    <li className={styles.item}>
-      <ButtonAddSampleBot onClick={openModal} icon={image}>
-        {template.title}
-      </ButtonAddSampleBot>
-      {isModalOpen && (
-        <ModalPopup onClick={closeModal}>
-          <BotTemplatePopup template={template} onClick={closeModal} />
-        </ModalPopup>
-      )}
-    </li>
-  );
-};
+import Template from '../template/template';
 
 const Templates: FC = () => {
   const { templatesBots } = useAppSelector(getTemplatesBotsSel);
